Rewrite to a library for import by other project
