+++ .babelrc
@@ -8,10 +8,11 @@
     }],
     "stage-2"
   ],
-  "plugins": ["transform-vue-jsx", "transform-runtime"],
+  "plugins": ["transform-runtime"],
   "env": {
     "test": {
-      "presets": ["env", "stage-2"]
+      "presets": ["env", "stage-2"],
+      "plugins": ["istanbul"]
     }
   }
 }
