+++ config/dev.env.js
@@ -1,6 +1,5 @@
-'use strict'
-const merge = require('webpack-merge')
-const prodEnv = require('./prod.env')
+var merge = require('webpack-merge')
+var prodEnv = require('./prod.env')
 
 module.exports = merge(prodEnv, {
   NODE_ENV: '"development"'
