+++ config/prod.env.js
@@ -1,4 +1,3 @@
-'use strict'
 module.exports = {
   NODE_ENV: '"production"'
 }
