+++ index.html
@@ -1,12 +1,1 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <title>vue-draggable-progressbar</title>
-  </head>
-  <body>
-    <div id="app"></div>
-    <!-- built files will be auto injected -->
-  </body>
-</html>
+<!-- This file will be overwritten and is used to display your docs. Instead, the index-template.html file contains the template used to generate this file. -->
